Разделил методы PUT / PATCH/ DELETE по файлам
